Fix the trigger for the chmod command
The previous trigger is just a string constant. It would run during the
first apply but would be skipped when terraform is applied again because
there was no change to the trigger. This changes the trigger to the
SHA256 hash of the private key so that chmod will be rerun when the file
is recreated.